skeleton for faces clock
The idea is to have a less complex digital clock that can load different
faces, instead of trying to handle every single use-case in just one
implementation. As discussed during the Plasma sprint.

This is a copy of just the QML bits of the digital-clock, with a
different plugin name.