Use a separate config value for Wayland font DPI
Summary:
On X our font DPI is used in addition to the scaling.
i.e I want scale 2 and font DPI at 192 (96*2)

On Wayland it's the opposite. The font is scaled *with* the output
scaling. So I want it at 96 (or unset)

Using the same settings makes swithing to and from X and wayland hard,
and also makes the first Wayland run suck.

There are potentially some users who want it 110 at both, so we do still
need the option (so this isn't a perfect solution)

Can we just divide it? Not really as we don't know what scale the user
will be using and it might not just be one value.

I'll make the Font KCM write to two different config values depending on
where we're currently running.

Reviewers: #plasma

Subscribers: plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D7830